Conflicts:
	plugins/hosts/backend.py
	plugins/lib_rcconf/api.py
	plugins/services/__init__.py

+++ plugins/terminal/main.py
@@ -6,10 +6,19 @@ import os
 import threading
 import shlex
 import signal
+import json
+import gzip
+import StringIO
 from base64 import b64decode, b64encode
 
 from twisted.internet import protocol
 from twisted.internet import reactor
+
+import pyte
+
+
+TERM_W = 120
+TERM_H = 30
 
 
 class TerminalPlugin(CategoryPlugin, URLHandler):
@@ -19,6 +28,7 @@ class TerminalPlugin(CategoryPlugin, URLHandler):
 
     def on_session_start(self):
         self._proc = None
+        self._inputting = False
         
     def on_init(self):
         if self._proc is None:
@@ -26,12 +36,20 @@ class TerminalPlugin(CategoryPlugin, URLHandler):
     
     def get_ui(self):
         ui = self.app.inflate('terminal:main')
+        if self._inputting:
+            ui.append('main', UI.CodeInputBox(
+                id='dlgBatch',
+                value='',
+                text='Batch input:',
+            ))
         return ui
 
     def start(self):
         self._proc = PTYProtocol()
         env = os.environ
-        env['TERM'] = 'vt100'
+        env['TERM'] = 'linux'
+        env['COLUMNS'] = str(TERM_W)
+        env['LINES'] = str(TERM_H)
         sh = self.app.get_config(self).shell
         reactor.spawnProcess(
             self._proc, 
@@ -54,7 +72,11 @@ class TerminalPlugin(CategoryPlugin, URLHandler):
             data = self._proc.history()
         else:   
             data = self._proc.read()
-        return b64encode(data)
+        sio = StringIO.StringIO()
+        gz = gzip.GzipFile(fileobj=sio, mode='w')
+        gz.write(json.dumps(data))
+        gz.close()
+        return b64encode(sio.getvalue())
 
     @url('^/term-post/.+')
     def post(self, req, start_response):
@@ -64,15 +86,26 @@ class TerminalPlugin(CategoryPlugin, URLHandler):
     
     @event('button/click')
     def click(self, evt, params, vars):
-        self.restart()
+        if params[0] == 'restart':
+            self.restart()
+        if params[0] == 'input':
+            self._inputting = True
+            
+    @event('dialog/submit')
+    def submit(self, evt, params, vars=None):
+        if vars.getvalue('action', None) == 'OK':
+            self._proc.write(vars.getvalue('value', ''))
+        self._inputting = False
         
     
 class PTYProtocol(protocol.ProcessProtocol):
     def __init__(self):
         self.data = ''
-        self.hist = ''
         self.lock = threading.Lock()
         self.cond = threading.Condition()
+        self.term = pyte.DiffScreen(TERM_W,TERM_H)
+        self.stream = pyte.Stream()
+        self.stream.attach(self.term)
         
     def connectionMade(self):
         pass
@@ -81,25 +114,40 @@ class PTYProtocol(protocol.ProcessProtocol):
         with self.cond:
             with self.lock:
                 self.data += data
-                self.hist += data
-                self.hist = self.hist[-2000:]
             self.cond.notifyAll()        
 
     def read(self):
         with self.cond:
             if len(self.data) == 0:
-                self.cond.wait(5)
+                self.cond.wait(15)
             with self.lock:
                 data = self.data
                 self.data = ''
-        return data
+                if len(data) > 0:
+                    self.stream.feed(unicode(str(data)))
+        return self.format()
         
     def history(self):
-        return self.hist
+        return self.format(full=True)
+        
+    def format(self, full=False):
+        l = {}
+        self.term.dirty.add(self.term.cursor.y)
+        for k in self.term.dirty:
+            l[k] = self.term[k]
+        self.term.dirty.clear()
+        r = {
+            'lines': self.term if full else l,
+            'cx': self.term.cursor.x,
+            'cy': self.term.cursor.y,
+            'cursor': not self.term.cursor.hidden,
+        }
+        return r
         
     def write(self, data):
         self.transport.write(data)
 
     def kill(self):
-        os.kill(self.transport.pid, signal.SIGKILL)
+        if self.transport.pid:
+            os.kill(self.transport.pid, signal.SIGKILL)
